slit dbHashPool and PoolCollection file
